Took out cura and slic3r mapping classes
improved fan settings
put in enable disable fan
